Move types/functions to cardano-testnet
