+++ cardano-node/src/Cardano/Node/Configuration/Topology.hs
@@ -11,7 +11,6 @@ module Cardano.Node.Configuration.Topology
   , NodeHostIPv4Address(..)
   , NodeHostIPv6Address(..)
   , NodeSetup(..)
-  , NodeId(..)
   , RemoteAddress(..)
   , nodeAddressToSockAddr
   , readTopologyFile
@@ -28,7 +27,6 @@ import           Ouroboros.Consensus.Util.Condense (Condense (..))
 import           Control.Exception (Exception (..), IOException)
 import qualified Control.Exception as Exception
 import           Data.Aeson
-import           Data.Aeson.Types (parseFail)
 import           Data.Bifunctor (Bifunctor (..))
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy.Char8 as LBS
@@ -58,19 +56,6 @@ data RemoteAddress = RemoteAddress
   -- if an IP address is given valency is used as
   -- a Boolean value, @0@ means to ignore the address;
   } deriving (Eq, Ord, Show)
-
-newtype NodeId = NodeId Int
-  deriving (Eq, Ord, Show)
-
-instance ToJSON NodeId where
-  toJSON (NodeId i) = String $ Text.pack $ "node_" ++ show i
-
-instance FromJSON NodeId where
-  parseJSON = withText "NodeId" $ \t -> case Text.breakOn "_" t of
-    ("node", textId) -> case eitherDecodeStrictText textId of
-      Right i -> pure $ NodeId i
-      Left _ -> parseFail $ "Incorrect format for NodeId: " ++ show t
-    _ -> parseFail $ "Incorrect format for NodeId: " ++ show t
 
 -- | Parse 'raAddress' field as an IP address; if it parses and the valency is
 -- non zero return corresponding NodeAddress.
